+++ neurocaps/__init__.py
@@ -4,4 +4,4 @@ from . import analysis, extraction, exceptions
 __all__ = ["analysis", "extraction", "exceptions"]
 
 # Version in single place
-__version__ = "0.24.8"
+__version__ = "0.25.0"