+++ neurocaps/_utils/__init__.py
@@ -11,5 +11,5 @@ from .analysis import (
     _run_kmeans,
     _save_contents,
 )
-from .extraction import _TimeseriesExtractorGetter, _check_confound_names, _extract_timeseries
+from .extraction import _TimeseriesExtractorGetter, _check_confound_names, _extract_timeseries, _standardize
 from .plotting_defaults import _PlotDefaults