+++ neurocaps/_utils/extraction/__init__.py
@@ -1,3 +1,3 @@
 from .check_confound_names import _check_confound_names
-from .extract_timeseries import _extract_timeseries
+from .extract_timeseries import _extract_timeseries, _standardize
 from .timeseriesextractorgetter import _TimeseriesExtractorGetter