+++ tests/test_CAP.py
@@ -127,8 +127,8 @@ def test_groups_without_cluster_selection(standardize, runs):
         assert np.array_equal(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"])
         assert np.array_equal(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"])
     else:
-        assert np.allclose(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"])
-        assert np.allclose(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"])
+        assert np.allclose(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"], atol=0.00001)
+        assert np.allclose(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"], atol=0.00001)
 
     # Demonstrates that label assignment using `.predict` method and paired with proper subject level standardization
     # using the mean and std dev computed from the concatenated data produces the same labels stored in `.labels_`
@@ -151,8 +151,8 @@ def test_groups_without_cluster_selection(standardize, runs):
         assert np.array_equal(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"])
         assert np.array_equal(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"])
     else:
-        assert np.allclose(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"])
-        assert np.allclose(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"])
+        assert np.allclose(cap_analysis.concatenated_timeseries["A"], concatenated_timeseries["A"], atol=0.00001)
+        assert np.allclose(cap_analysis.concatenated_timeseries["B"], concatenated_timeseries["B"], atol=0.00001)
 
     labels = predict_labels(timeseries, cap_analysis, standardize, "A", runs=run_nums)
     assert np.array_equal(labels, cap_analysis.kmeans["A"].labels_)